Move CellIDTranslator into a new file
